Make grayscale input be 16-bit

+++ NetPimpl.h
@@ -9,7 +9,7 @@ namespace NetPimpl
 {
 #ifdef DLIB_DNN_PIMPL_WRAPPER_GRAYSCALE_INPUT
     // TODO: use definition from MemoryManager.h
-    typedef dlib::matrix<uint8_t,0,0,dlib::memory_manager_stateless<uint8_t>::kernel_2_3e> image_type;
+    typedef dlib::matrix<uint16_t,0,0,dlib::memory_manager_stateless<uint16_t>::kernel_2_3e> image_type;
 #else
     typedef dlib::matrix<dlib::rgb_pixel,0,0,dlib::memory_manager_stateless<uint8_t>::kernel_2_3e> image_type;
 #endif
